pyproject.toml
[project]
name = "nancy-brain"
version = "0.1.5"
description = "Turn any GitHub repository into a searchable knowledge base for AI agents"
readme = "README.md"
requires-python = ">=3.10"
dynamic = ["authors"]   # <-- this makes Hatch load your metadata hook
license = {text = "MIT"}
keywords = ["AI", "RAG", "knowledge-base", "search", "embeddings", "research", "github"]
dependencies = [
    "aiohttp>=3.8.0",
    "slack-sdk>=3.18.0",
    "slack-bolt>=1.18.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=5.3",
    "docutils>=0.20",
    "pylatexenc>=2.10",
    "txtai>=7.0.0",
    "requests>=2.28.0",
    "torch>=1.12.1",
    "faiss-cpu>=1.7.1.post2",
    "transformers>=4.45.0",
    "regex>=2022.8.17",
    "nb4llm>=0.1.3",
    "pydantic>=2.0.0",   
    "mcp>=1.13.0",       
    "click>=8.0.0",      
    "streamlit>=1.28.0", 
    "fastapi>=0.110.0",  
    "uvicorn>=0.29.0",   
    "passlib[bcrypt]>=1.7.4", 
    "python-jose>=3.3.0"  
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Indexing",
    "Environment :: Console",
    "Environment :: Web Environment",
]

[project.urls]
Homepage = "https://github.com/AmberLee2427/nancy-brain"
Documentation = "https://github.com/AmberLee2427/nancy-brain/blob/main/README.md"
Repository = "https://github.com/AmberLee2427/nancy-brain"
Issues = "https://github.com/AmberLee2427/nancy-brain/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=25.1.0",
    "flake8>=7.3.0",
    "coverage>=7.0.0",
    "pre-commit>=3.0.0",
    "ipykernel",
    "jupyter",
    "ipywidgets",
    "build>=0.10.0",
    "twine>=4.0.0",
    "bump-my-version>=0.11.0",
    "hatch"
]
pdf = [
    "tika>=2.6.0",
    "PyPDF2",
    "pdfplumber",
    "pymupdf"
]
docs = [
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.20.0",
    "mkdocs-autorefs>=0.4.1",
    "mkdocs-click>=0.8.0"
]

[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "nancy_brain",
    "rag_core",
    "connectors",
    "scripts"
]

[tool.hatch.build]
include = [
  "hatch_hooks.py",
  "nancy_brain/**",
  "rag_core/**",
  "connectors/**",
  "scripts/**",
]

[project.scripts]
nancy-brain = "nancy_brain.cli:cli"

[tool.hatch.metadata.hooks.custom]
path = "hatch_hooks.py"

[tool.bumpversion]
current_version = "0.1.5"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
tag = true
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = true
message = "Bump version: {current_version} → {new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "nancy_brain/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[[tool.bumpversion.files]]
filename = "rag_core/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'
